pyproject.toml
[project]
name = "adomcore"
version = "0.1.0"
description = "Single-instance agent runtime with file-based persistence"
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "apscheduler>=3.11",
    "typer>=0.15",
    "loguru>=0.7",
    "pyyaml>=6.0",
    "pyjson5>=1.6",
    "aiofiles>=24.1",
    "anthropic>=0.49",
    "openai>=1.70",
    "tiktoken>=0.9",
    "atomic-agents>=1.0",
    "python-multipart>=0.0.20",
    "anyio>=4.9",
]

[project.scripts]
adom = "adomcore.cli:app"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"
include = ["src"]
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff.lint]
select = ["I", "F401", "UP"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyright[nodejs]>=1.1.408",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
]
